requirements.txt
langchain
langchain-openai
openai
pandas
openpyxl
sqlalchemy
pydantic
python-dotenv
faiss-cpu
langchain-text-splitters